move hook logic into price badges
